requirements.txt
# Web Framework
streamlit==1.32.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# API and Data Validation
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2 